% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SongjoowhanToolsRfunctions.R
\name{func8}
\alias{func8}
\title{Computes matrix}
\usage{
func8(a, x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar
}
\description{
Computes matrix
}
\examples{
func8(a,x)

}
