% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SongjoowhanToolsRfunctions.R
\name{func9}
\alias{func9}
\title{Standardizes matrix}
\usage{
func9(x)
}
\arguments{
\item{x}{vector}
}
\value{
scalar
}
\description{
Standardizes matrix
}
\examples{
func9(a)

}
